+++ lib/src/constants/index.dart
@@ -7,3 +7,5 @@ export 'outbound_packet_encoding.dart';
 export 'clash_mode.dart';
 export 'flutter_sing_box_constants.dart';
 export 'rule_action.dart';
+export 'windows_service.dart';
+export 'inbound_type.dart';